Refactored binary file into seperate modules
* Refactored the main binary source file into seperate modules.
* TODO: Move useful bits that aren't directly cli related into the
  lib directory.
