+++ lib/src/bin/phylum-cli/commands.rs
@@ -1,0 +1,5 @@
+pub mod auth;
+pub mod jobs;
+pub mod lock_files;
+pub mod packages;
+pub mod projects;
